[WIP] Extract custom hooks from controllers
