+++ packages/ra-core/src/controller/ListController.spec.tsx
@@ -5,7 +5,6 @@ import TextField from '@material-ui/core/TextField/TextField';
 
 import ListController, {
     getListControllerProps,
-    getQuery,
     sanitizeListRestProps,
 } from './ListController';
 
@@ -224,161 +223,6 @@ describe('ListController', () => {
             ).toEqual({
                 foo: 1,
                 bar: 'hello',
-            });
-        });
-    });
-    describe('getQuery', () => {
-        it('Returns the values from the location first', () => {
-            const query = getQuery({
-                location: {
-                    search: `?page=3&perPage=15&sort=name&order=ASC&filter=${JSON.stringify(
-                        { name: 'marmelab' }
-                    )}`,
-                },
-                params: {
-                    page: 1,
-                    perPage: 10,
-                    sort: 'city',
-                    order: 'DESC',
-                    filter: {
-                        city: 'Dijon',
-                    },
-                },
-                filterDefaultValues: {},
-                perPage: 50,
-                sort: {
-                    field: 'company',
-                    order: 'DESC',
-                },
-            });
-
-            expect(query).toEqual({
-                page: '3',
-                perPage: '15',
-                sort: 'name',
-                order: 'ASC',
-                filter: {
-                    name: 'marmelab',
-                },
-            });
-        });
-        it('Extends the values from the location with those from the props', () => {
-            const query = getQuery({
-                location: {
-                    search: `?filter=${JSON.stringify({ name: 'marmelab' })}`,
-                },
-                params: {
-                    page: 1,
-                    perPage: 10,
-                    sort: 'city',
-                    order: 'DESC',
-                    filter: {
-                        city: 'Dijon',
-                    },
-                },
-                filterDefaultValues: {},
-                perPage: 50,
-                sort: {
-                    field: 'company',
-                    order: 'DESC',
-                },
-            });
-
-            expect(query).toEqual({
-                page: 1,
-                perPage: 50,
-                sort: 'company',
-                order: 'DESC',
-                filter: {
-                    name: 'marmelab',
-                },
-            });
-        });
-        it('Sets the values from the redux store if location does not have them', () => {
-            const query = getQuery({
-                location: {
-                    search: ``,
-                },
-                params: {
-                    page: 2,
-                    perPage: 10,
-                    sort: 'city',
-                    order: 'DESC',
-                    filter: {
-                        city: 'Dijon',
-                    },
-                },
-                filterDefaultValues: {},
-                perPage: 50,
-                sort: {
-                    field: 'company',
-                    order: 'DESC',
-                },
-            });
-
-            expect(query).toEqual({
-                page: 2,
-                perPage: 10,
-                sort: 'city',
-                order: 'DESC',
-                filter: {
-                    city: 'Dijon',
-                },
-            });
-        });
-        it('Extends the values from the redux store with those from the props', () => {
-            const query = getQuery({
-                location: {
-                    search: ``,
-                },
-                params: {
-                    page: 2,
-                    sort: 'city',
-                    order: 'DESC',
-                    filter: {
-                        city: 'Dijon',
-                    },
-                },
-                filterDefaultValues: {},
-                perPage: 50,
-                sort: {
-                    field: 'company',
-                    order: 'DESC',
-                },
-            });
-
-            expect(query).toEqual({
-                page: 2,
-                perPage: 50,
-                sort: 'city',
-                order: 'DESC',
-                filter: {
-                    city: 'Dijon',
-                },
-            });
-        });
-        it('Uses the filterDefaultValues if neither the location or the redux store have them', () => {
-            const query = getQuery({
-                location: {
-                    search: ``,
-                },
-                params: {},
-                filterDefaultValues: { city: 'Nancy' },
-                perPage: 50,
-                sort: {
-                    field: 'company',
-                    order: 'DESC',
-                },
-            });
-
-            expect(query).toEqual({
-                page: 1,
-                perPage: 50,
-                sort: 'company',
-                order: 'DESC',
-                filter: {
-                    city: 'Nancy',
-                },
             });
         });
     });
