+++ packages/ra-core/src/controller/useList.spec.ts
@@ -1,0 +1,163 @@
+import { getQuery } from './useList';
+import {
+    SORT_DESC,
+    SORT_ASC,
+} from '../reducer/admin/resource/list/queryReducer';
+
+describe('ListController', () => {
+    describe('getQuery', () => {
+        it('Returns the values from the location first', () => {
+            const query = getQuery({
+                location: {
+                    search: `?page=3&perPage=15&sort=name&order=ASC&filter=${JSON.stringify(
+                        { name: 'marmelab' }
+                    )}`,
+                },
+                params: {
+                    page: 1,
+                    perPage: 10,
+                    sort: 'city',
+                    order: SORT_DESC,
+                    filter: {
+                        city: 'Dijon',
+                    },
+                },
+                filterDefaultValues: {},
+                perPage: 50,
+                sort: {
+                    field: 'company',
+                    order: SORT_DESC,
+                },
+            });
+
+            expect(query).toEqual({
+                page: 3,
+                perPage: 15,
+                sort: 'name',
+                order: SORT_ASC,
+                filter: {
+                    name: 'marmelab',
+                },
+            });
+        });
+        it('Extends the values from the location with those from the props', () => {
+            const query = getQuery({
+                location: {
+                    search: `?filter=${JSON.stringify({ name: 'marmelab' })}`,
+                },
+                params: {
+                    page: 1,
+                    perPage: 10,
+                    sort: 'city',
+                    order: SORT_DESC,
+                    filter: {
+                        city: 'Dijon',
+                    },
+                },
+                filterDefaultValues: {},
+                perPage: 50,
+                sort: {
+                    field: 'company',
+                    order: SORT_DESC,
+                },
+            });
+
+            expect(query).toEqual({
+                page: 1,
+                perPage: 50,
+                sort: 'company',
+                order: SORT_DESC,
+                filter: {
+                    name: 'marmelab',
+                },
+            });
+        });
+        it('Sets the values from the redux store if location does not have them', () => {
+            const query = getQuery({
+                location: {
+                    search: ``,
+                },
+                params: {
+                    page: 2,
+                    perPage: 10,
+                    sort: 'city',
+                    order: SORT_DESC,
+                    filter: {
+                        city: 'Dijon',
+                    },
+                },
+                filterDefaultValues: {},
+                perPage: 50,
+                sort: {
+                    field: 'company',
+                    order: SORT_DESC,
+                },
+            });
+
+            expect(query).toEqual({
+                page: 2,
+                perPage: 10,
+                sort: 'city',
+                order: SORT_DESC,
+                filter: {
+                    city: 'Dijon',
+                },
+            });
+        });
+        it('Extends the values from the redux store with those from the props', () => {
+            const query = getQuery({
+                location: {
+                    search: ``,
+                },
+                params: {
+                    page: 2,
+                    sort: 'city',
+                    order: SORT_DESC,
+                    filter: {
+                        city: 'Dijon',
+                    },
+                },
+                filterDefaultValues: {},
+                perPage: 50,
+                sort: {
+                    field: 'company',
+                    order: SORT_DESC,
+                },
+            });
+
+            expect(query).toEqual({
+                page: 2,
+                perPage: 50,
+                sort: 'city',
+                order: SORT_DESC,
+                filter: {
+                    city: 'Dijon',
+                },
+            });
+        });
+        it('Uses the filterDefaultValues if neither the location or the redux store have them', () => {
+            const query = getQuery({
+                location: {
+                    search: ``,
+                },
+                params: {},
+                filterDefaultValues: { city: 'Nancy' },
+                perPage: 50,
+                sort: {
+                    field: 'company',
+                    order: SORT_DESC,
+                },
+            });
+
+            expect(query).toEqual({
+                page: 1,
+                perPage: 50,
+                sort: 'company',
+                order: SORT_DESC,
+                filter: {
+                    city: 'Nancy',
+                },
+            });
+        });
+    });
+});
